feat: add file utils

+++ pipeline.py
@@ -1,7 +1,6 @@
 # MODULES
 import time
 import os
-import shutil
 from pathlib import Path
 
 # MODELS
@@ -11,80 +10,59 @@ from models.config import Config
 from libs.clustering_lib import Clustering
 
 # UTILS
-from utils import mailing
+from utils import mailing, file
 
-# CONFIGS
+# CORE
 from core import LOGGER, CONFIGS
 
 
 class PipeLine:
     def __init__(self, config: Config) -> None:
         self.config = config
+        self.clustering = Clustering(config=self.config.clustering)
 
     def start(self):
-        clustering = Clustering(config=self.config)
         while True:
-            time.sleep(5)
+            time.sleep(self.config.time_sleep)
 
-            klarf_paths = sorted(
-                [
-                    Path(os.path.join(self.config.path.input, f))
-                    for f in os.listdir(self.config.path.input)
-                    if os.path.isfile(os.path.join(self.config.path.input, f))
-                ],
-                key=lambda x: os.path.getmtime(os.path.join(self.config.path.input, x)),
+            klarf_paths, nbr_klarfs = file.get_files(
+                path=self.config.path.input, sort_by_modification_date=True
             )
-
-            nbr_klarfs = len(klarf_paths)
 
             if not nbr_klarfs == 0:
                 for klarf_path in klarf_paths:
                     klarf = os.path.basename(klarf_path)
+                    klarf_name, klarf_extension = os.path.splitext(klarf)
 
                     try:
-                        size = os.path.getsize(klarf_path)
+                        if file.check_file_size(klarf_path):
+                            output_path = (
+                                Path(self.config.path.output)
+                                / f"{klarf_name}_clustered{klarf_extension}"
+                            )
 
-                        # wait until the file size stops changing
-                        while True:
-                            time.sleep(1)  # wait 1 second before checking again
-                            new_size = os.path.getsize(klarf_path)
-                            if new_size == size:
-                                break
-                            size = new_size
+                            results = self.clustering.apply(
+                                klarf_path=klarf_path,
+                                output_path=output_path,
+                            )
 
-                        clustering.apply(
-                            klarf_path=klarf_path,
-                            baby_klarf=True,
-                        )
-
-                        if os.path.exists(klarf_path):
-                            os.remove(klarf_path)
+                            if not len(results) == 0 and os.path.exists(klarf_path):
+                                os.remove(klarf_path)
+                            else:
+                                LOGGER.error(
+                                    msg=f"Unable to remove {klarf=}",
+                                )
 
                     except Exception as ex:
-                        shutil.move(
+                        file.move(
                             src=klarf_path,
-                            dst=os.path.join(
-                                self.config.path.error, os.path.basename(klarf_path)
-                            ),
+                            dst=os.path.join(self.config.path.error, klarf),
                         )
 
-                        message_error = f"{klarf=} processing failed, moved to {self.config.path.error}"
-
-                        html = f"""\
-                            <html>
-                                <body>
-                                    <p>{message_error}</p>
-                                </body>
-                            </html>
-                        """
-
-                        mailing.send_mail(
-                            host=CONFIGS.mailing.host,
-                            port=CONFIGS.mailing.port,
-                            sender=CONFIGS.mailing.sender,
-                            receiver=CONFIGS.mailing.receiver,
-                            subject=f"Clustering - Error on {klarf}",
-                            msg_html=html,
+                        message_error = mailing.send_mail_error(
+                            klarf=klarf,
+                            error_path=self.config.path.error,
+                            config=CONFIGS.mailing,
                         )
 
                         LOGGER.critical(
